requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
affine=2.3.1=pypi_0
appdirs=1.4.4=pypi_0
asciitree=0.3.3=pypi_0
attrs=21.4.0=pypi_0
basemap=1.2.2=py39h7980778_5
basemap-data=1.3.2=pyhd8ed1ab_1
basemap-data-hires=1.3.2=pyhd8ed1ab_1
beautifulsoup4=4.11.1=pypi_0
bokeh=2.4.2=pypi_0
bottleneck=1.3.4=pypi_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.5.18.1=ha878542_0
cdsapi=0.5.1=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=pypi_0
cfgrib=0.9.10.1=pypi_0
cftime=1.6.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.1.2=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=2.0.0=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
dask=2022.4.1=pypi_0
distributed=2022.4.1=pypi_0
docopt=0.6.2=pypi_0
eccodes=1.4.1=pypi_0
et-xmlfile=1.1.0=pypi_0
fasteners=0.17.3=pypi_0
findlibs=0.0.2=pypi_0
fonttools=4.33.3=py39hb9d737c_0
freetype=2.10.4=h0708190_1
fsspec=2022.3.0=pypi_0
gdal=3.3.2=pypi_0
geos=3.10.2=h9c3ff4c_0
giflib=5.2.1=h36c2ea0_2
h5netcdf=1.0.0=pypi_0
h5py=3.6.0=pypi_0
heapdict=1.0.1=pypi_0
idna=3.3=pypi_0
install=1.3.5=pypi_0
jinja2=3.1.1=pypi_0
jpeg=9e=h166bdaf_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py39hf939315_1
krb5=1.19.3=h3790be6_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=15_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=15_linux64_openblas
libcurl=7.83.1=h7bff187_0
libdeflate=1.12=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=h58526e2_2
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
liblapack=3.9.0=15_linux64_openblas
libnghttp2=1.47.0=h727a467_0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.37=h21135ba_2
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.4.0=hc85c160_1
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libzlib=1.2.12=h166bdaf_0
llvmlite=0.38.0=pypi_0
locket=1.0.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.1=pypi_0
matplotlib-base=3.5.2=py39h700656a_0
msgpack=1.0.3=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nc-time-axis=1.4.1=pypi_0
ncurses=6.3=h27087fc_1
netcdf4=1.5.8=pypi_0
numba=0.55.1=pypi_0
numbagg=0.2.1=pypi_0
numcodecs=0.9.1=pypi_0
numpy=1.21.6=pypi_0
openjpeg=2.4.0=hb52868f_1
openpyxl=3.0.9=pypi_0
openssl=1.1.1o=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=pypi_0
partd=1.2.0=pypi_0
pillow=9.1.0=pypi_0
pip=22.1.2=pyhd8ed1ab_0
pooch=1.6.0=pypi_0
proj=8.2.1=h277dcde_0
psutil=5.9.0=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pypi_0
pydap=3.2.2=pypi_0
pyparsing=3.0.7=pypi_0
pyproj=3.3.0=py39hab5ddba_1
pyshp=2.3.0=pyhd8ed1ab_0
python=3.9.0=hffdb5ce_5_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
rasterio=1.2.10=pypi_0
readline=8.1.2=h0f457ee_0
requests=2.27.1=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.11.2=pypi_0
setuptools=62.3.4=py39hf3d152e_0
six=1.16.0=pyh6c4a22f_0
snuggs=1.4.7=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.5=h4ff8645_0
tblib=1.7.0=pypi_0
tk=8.6.12=h27826a3_0
toolz=0.11.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.63.0=pypi_0
typing-extensions=4.2.0=pypi_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39hb9d737c_1
urllib3=1.26.9=pypi_0
webob=1.8.7=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xarray=2022.3.0=pypi_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zarr=2.11.3=pypi_0
zict=2.1.0=pypi_0
zlib=1.2.12=h166bdaf_0
zstd=1.5.2=h8a70e8d_1
